Estimation of error fraction in SSO and client.

A shared service organization, SSO, performs transactions for various clients. Each transaction can potential be wrong, or partially wrong. To get an estimate of the error rate of all the transactions we can take a random sample of the transactions from the SSO. This random sample is defined by the number of transactions, n_SSO, in the sample, and the sum of partial and full errors found, k_SSO. Given the number of transactions for a specific client, N_client, we can then estimate the error rate for that specific client. This function makes such an estimation, using the assumption that we can consider the transactions of the client to be a random sample of all the transactions performed by the SSO. The function returns the probability graph of the error fraction in the SSO, and the probability graph of the error fraction in the client transactions For the estimation, the function uses a two dimensional grid of postulated values for the error fractions of SSO, and client. Both dimensions of the grid have the same granularity: S; this makes it easier to compare the two resulting probability curves.

SSO_estimate(k_SSO = 0, n_SSO = 350, N_client = 1000, S = 2000)

Arguments

k_SSO

The sum of partial and full errors found in the sample from the SSO. k_SSO is a non negative real number. Default 0.

n_SSO

The number of transactions in the sample from the SSO. n_SSO is a positive integer. Note that we follow the convention to use a lower case n for the number of transactions in the sample (in this case from the SSO transactions). Default 350.

N_client

The number of transactions of the client. N_client is a positive integer. Note that we follow the convention to use an upper case N for the number of transactions in the mass of all transactions (in this case of the client). Default 1000.

S

The number of segments, represented by their midpoints, in which is divided each of the dimensions of the two dimensional grid used for computing the probability graphs of the error fraction of the transactions in the SSO, and the errors in the transactions of the client. So S is the granularity of the grid. S is a positive integer. Default 2000.

Value

A list of named elements containing:

    - P_SSO, the probability graph of the error fraction in SSO
    - P_client, the probability graph of the error fraction in the client transactions
    - the input parameters

Examples

y <- SSO_estimate(k_SSO = 1, n_SSO = 3000, N_client = 500, S = 501)