ssoestimate: Estimate Error Rate of Services for Client of SSO

Estimate maximum error rate of service provided to any client of a Shared Service Organization (SSO) by using a sufficiently large, but relatively small, sample of the overall population. The key assumption for this is that the service for each client can be considered as a random sample from the service of the overall population.

See also

Author

Maintainer: Crist-Jan Doedens cfjdoedens@chance-matters.nl (ORCID)